requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.08.22=hca03da5_0
gurobi=10.0.3=py311_0
libffi=3.4.4=hca03da5_0
ncurses=6.4=h313beb8_0
openssl=3.0.11=h1a28f6b_2
pip=23.3=py311hca03da5_0
python=3.11.5=hb885b13_0
readline=8.2=h1a28f6b_0
setuptools=68.0.0=py311hca03da5_0
sqlite=3.41.2=h80987f9_0
tk=8.6.12=hb8d0fd4_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py311hca03da5_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
